Do not decrement sysRef of allocation from rsGetAllocation()
Bug: 31060926

rsGetAllocation() is a deprecated API that leaves sysRef as is (0 for
new allocations.) This breaks assumptions of the latest llvm-rs-cc in
build tools 24.0.1, which relies on the invariant that any RenderScript
runtime API that returns a RS object will have its sysRef set to 1.

This CL skips rsGetAllocation() calls and does not generate an extra
rsClearObject() call to decrement the sysRef of its returned object.

Change-Id: I565e268cee0ba56985590a0343094fb9a7f12915
Signed-off-by: mydongistiny <jaysonedson@gmail.com>